Add documentation on stopping and deleting apps
The commands related to stopping and deleting apps are simple, but
we want to convey the consequences and things to consider when
doing so. This will allow Cloud Foundry users to keep their orgs and
spaces tidy, and avoid errors, such as routes already existing because
they were initially created in another space and were never deleted.